Improve support for app.trustedHosts values.
